major refactoring of packages
