+++ ibaqpy/commands/features2peptides.py
@@ -1,5 +1,5 @@
 import click
-from ibaqpy.bin.peptide_normalization import peptide_normalization
+from ibaqpy.ibaq.peptide_normalization import peptide_normalization
 
 
 @click.command("features2peptides", short_help="Convert features to parquet file.")
