+++ ibaqpy/commands/merge_condition_files.py
@@ -2,7 +2,7 @@ import os
 import click
 import pandas as pd
 
-from ibaqpy.bin.ibaqpy_commons import CONDITION
+from ibaqpy.ibaq.ibaqpy_commons import CONDITION
 
 
 @click.command()
