+++ ibaqpy/commands/tsne_visualization.py
@@ -8,7 +8,7 @@ import seaborn as sns
 from sklearn.decomposition import PCA
 from sklearn.manifold import TSNE
 
-from ibaqpy.bin.ibaqpy_commons import PROTEIN_NAME, SAMPLE_ID, IBAQ_LOG
+from ibaqpy.ibaq.ibaqpy_commons import PROTEIN_NAME, SAMPLE_ID, IBAQ_LOG
 
 
 # function to compute principal components
